PIZ11: Refactor Repositories
