+++ app/receivers/ingredient.py
@@ -1,5 +1,5 @@
 from .base import BaseReceiver
-from ..repositories.managers import IngredientManager
+from ..repositories.managers.ingredient import IngredientManager
 
 
 class IngredientReceiver(BaseReceiver):
