--- app/repositories/managers.py
@@ -1,163 +1,0 @@
-from datetime import datetime
-from typing import Any, List, Optional, Sequence
-
-from sqlalchemy.sql import text, column
-
-from .models import Beverage, Ingredient, Order, OrderDetail, Size, db
-from .serializers import (
-    BeverageSerializer,
-    IngredientSerializer,
-    OrderSerializer,
-    SizeSerializer,
-    ma
-)
-
-
-class BaseManager:
-    model: Optional[db.Model] = None
-    serializer: Optional[ma.SQLAlchemyAutoSchema] = None
-    session = db.session
-
-    @classmethod
-    def get_all(cls):
-        serializer = cls.serializer(many=True)
-        _objects = cls.model.query.all()
-        result = serializer.dump(_objects)
-        return result
-
-    @classmethod
-    def get_by_id(cls, _id: Any):
-        entry = cls.model.query.get(_id)
-        return cls.serializer().dump(entry)
-
-    @classmethod
-    def create(cls, entry: dict):
-        serializer = cls.serializer()
-        new_entry = serializer.load(entry)
-        cls.session.add(new_entry)
-        cls.session.commit()
-        return serializer.dump(new_entry)
-
-    @classmethod
-    def update(cls, _id: Any, new_values: dict):
-        cls.session.query(cls.model).filter_by(_id=_id).update(new_values)
-        cls.session.commit()
-        return cls.get_by_id(_id)
-
-
-class BeverageManager(BaseManager):
-    model = Beverage
-    serializer = BeverageSerializer
-
-    @classmethod
-    def get_by_id_list(cls, ids: Sequence):
-        return cls.session.query(cls.model).filter(cls.model._id.in_(set(ids))).all() or []
-
-
-class SizeManager(BaseManager):
-    model = Size
-    serializer = SizeSerializer
-
-
-class IngredientManager(BaseManager):
-    model = Ingredient
-    serializer = IngredientSerializer
-
-    @classmethod
-    def get_by_id_list(cls, ids: Sequence):
-        return cls.session.query(cls.model).filter(cls.model._id.in_(set(ids))).all() or []
-
-
-class OrderManager(BaseManager):
-    model = Order
-    serializer = OrderSerializer
-
-    @classmethod
-    def create(cls, order_data: dict, beverages: List[Beverage], ingredients: List[Ingredient]):
-        new_order = cls.model(**order_data)
-        cls.session.add(new_order)
-        cls.session.flush()
-        cls.session.refresh(new_order)
-        cls.session.add_all(
-            list(
-                map(
-                    lambda beverage, ingredient: OrderDetail(
-                        order_id=new_order._id,
-                        beverage_id=beverage._id,
-                        beverage_price=beverage.price,
-                        ingredient_id=ingredient._id,
-                        ingredient_price=ingredient.price
-                    ),
-                    beverages,
-                    ingredients
-                )
-            )
-        )
-        cls.session.commit()
-        return cls.serializer().dump(new_order)
-
-    @classmethod
-    def update(cls):
-        raise NotImplementedError(f'Method not suported for {cls.__name__}')
-
-
-class IndexManager(BaseManager):
-
-    @classmethod
-    def test_connection(cls):
-        cls.session.query(column('1')).from_statement(text('SELECT 1')).all()
-
-
-class ReportManager(BaseManager):
-    @classmethod
-    def get_report(cls):
-        return {
-            'most_requested_ingredient': cls.get_most_requested_ingredient(),
-            'month_with_more_revenue': cls.get_month_with_more_revenue(),
-            'best_customers': cls.get_best_customers(),
-        }
-
-    @classmethod
-    def get_most_requested_ingredient(cls):
-        most_requested_ingredient = cls.session.query(
-            OrderDetail.ingredient_id,
-            Ingredient.name,
-            db.func.count(OrderDetail.ingredient_id).label('num_requested')
-        ).join(Ingredient).group_by(OrderDetail.ingredient_id).order_by(text('num_requested DESC')).first()
-
-        return {'name': most_requested_ingredient.name, 'num_requested': most_requested_ingredient.num_requested}
-
-    @classmethod
-    def get_month_with_more_revenue(cls):
-        month_with_more_revenue = cls.session.query(
-            db.func.strftime('%m', Order.date).label('month'),
-            db.func.sum(Order.total_price).label('revenue')
-        ).group_by('month').order_by(text('revenue DESC')).first()
-
-        month_name = datetime.strptime(month_with_more_revenue.month, '%m').strftime('%B')
-
-        return {'month': month_name, 'revenue': round(month_with_more_revenue.revenue, 2)}
-
-    @classmethod
-    def get_best_customers(cls):
-        best_customers = cls.session.query(
-            Order.client_name,
-            Order.client_dni,
-            Order.client_address,
-            Order.client_phone,
-            db.func.count(Order._id).label('num_orders')
-        ) \
-        .group_by(Order.client_name, Order.client_dni, Order.client_address, Order.client_phone) \
-        .order_by(text('num_orders DESC')) \
-        .limit(3) \
-        .all()
-
-        best_customers = list(map(lambda client: {
-            "client_name": client[0],
-            "client_dni": client[1],
-            "client_address": client[2],
-            "client_phone": client[3],
-            "num_orders": client[4]
-        }, best_customers))
-
-        return {'customers': best_customers}
