+++ app/repositories/managers/__init__.py
@@ -1,0 +1,4 @@
+from .beverage import BeverageManager
+from .ingredient import IngredientManager
+from .order import OrderManager
+from .size import SizeManager
