+++ app/repositories/managers/beverage.py
@@ -1,0 +1,14 @@
+from typing import Sequence
+
+from .base import BaseManager
+from ..models import Beverage
+from ..serializers.beverage import BeverageSerializer
+
+
+class BeverageManager(BaseManager):
+    model = Beverage
+    serializer = BeverageSerializer
+
+    @classmethod
+    def get_by_id_list(cls, ids: Sequence):
+        return cls.session.query(cls.model).filter(cls.model._id.in_(set(ids))).all() or []
