+++ app/repositories/models/ingredient.py
@@ -1,0 +1,7 @@
+from app.plugins import db
+
+
+class Ingredient(db.Model):
+    _id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(80), nullable=False)
+    price = db.Column(db.Float, nullable=False)
