+++ app/repositories/serializers/beverage.py
@@ -1,0 +1,10 @@
+from app.plugins import ma
+from ..models import Beverage
+
+
+class BeverageSerializer(ma.SQLAlchemyAutoSchema):
+
+    class Meta:
+        model = Beverage
+        load_instance = True
+        fields = ('_id', 'name', 'price')
