+++ app/repositories/serializers/ingredient.py
@@ -1,0 +1,10 @@
+from app.plugins import ma
+from ..models import Ingredient
+
+
+class IngredientSerializer(ma.SQLAlchemyAutoSchema):
+
+    class Meta:
+        model = Ingredient
+        load_instance = True
+        fields = ('_id', 'name', 'price')
