+++ app/repositories/serializers/order.py
@@ -1,0 +1,24 @@
+from app.plugins import ma
+from ..models import Order
+from .size import SizeSerializer
+from .order_detail import OrderDetailSerializer
+
+
+class OrderSerializer(ma.SQLAlchemyAutoSchema):
+    size = ma.Nested(SizeSerializer)
+    detail = ma.Nested(OrderDetailSerializer, many=True)
+
+    class Meta:
+        model = Order
+        load_instance = True
+        fields = (
+            '_id',
+            'client_name',
+            'client_dni',
+            'client_address',
+            'client_phone',
+            'date',
+            'total_price',
+            'size',
+            'detail'
+        )
